requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.3
django-widget-tweaks==1.4.8
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.19
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
Werkzeug==2.0.1
WTForms==2.3.3
zope.interface==5.4.0
